Resolved conversations
- Changed file names to match directory structure
- Removed commented code
- Added padding to friends title
- Used hooks to fetch user and friends
